Probleme avec l'ajout d'une candidature

+++ src/main/java/models/Etudiant.java
@@ -47,7 +47,7 @@ public class Etudiant {
         return numeroEtudiant;
     }
 
-    public void setNumeroEtudiant(int numeroEtudiant) {
+    public void setNumeroEtudiant(double numeroEtudiant) {
         this.numeroEtudiant = numeroEtudiant;
     }
 
@@ -106,5 +106,3 @@ public class Etudiant {
         }
     }
 }
-//Pour les étudiants qui demandent une bourse, on veut connaitre leur nom, prénom, numéro
-//étudiant, note moyenne du dernier semestre validé.
